requirements.txt
aiogram==2.25.1
aiogram-calendar
python-dotenv==1.0.0
peewee==3.16.0
requests==2.28.2
pydantic==1.10.5
urllib3==1.26.14
asyncio==3.4.3
